Разделяет main.js на разные модули
