+++ src/util.js
@@ -1,0 +1,4 @@
+const render = (element, template, position = `beforeend`) =>
+  element.insertAdjacentHTML(position, template);
+
+export {render};
