+++ src/view/events-container.js
@@ -1,0 +1,4 @@
+const createEventsContainerTemplate = () =>
+  `<ul class="trip-days"></ul>`;
+
+export {createEventsContainerTemplate};
